[fix] False-positive error detection after FTP command execution
After executing an FTP command, netrw checks to see if the 1st line of
output from the command is blank. If not, it assumes an error has
occurred. This breaks on systems where non-interactive ftp echoes its
stdin to stdout, making netrw believe that an error has occurred every
time it performs an FTP operation.

Check whether ftp's output is identical to its input. If so, no error
has occurred.